+++ AhchaScholarship/.idea/misc.xml
@@ -3,43 +3,29 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable-v24/dday_textview_background_off_radius.xml" value="0.189" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/bnv_color_selector.xml" value="0.1605" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/dday_textview_background_radius.xml" value="0.189" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_access_time_24.xml" value="0.1605" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_alarm_off_24.xml" value="0.1675" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_calendar_today_24.xml" value="0.1605" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_category_24.xml" value="0.1675" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_emoji_events_24.xml" value="0.1675" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_filter_alt_24.xml" value="0.1675" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_home_24.xml" value="0.1605" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_launch_24.xml" value="0.189" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_school_24.xml" value="0.1605" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/ic_baseline_star_24.xml" value="0.1605" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/indicator_circle.xml" value="0.1675" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/round_border.xml" value="0.1605" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/drawable/square_border.xml" value="0.189" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/activity_detail_scholarship.xml" value="0.25" />
+        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/activity_detail_scholarship.xml" value="0.1" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/activity_main.xml" value="0.22" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/activity_outdoor_filter.xml" value="0.30932971014492755" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/favorite.xml" value="0.25" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/fragment_favorite.xml" value="0.2898550724637681" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/fragment_outdoor.xml" value="0.25" />
+        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/fragment_outdoor.xml" value="0.2898550724637681" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/fragment_scholarship.xml" value="0.33" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/item_outdoor.xml" value="0.30932971014492755" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/item_scholarship.xml" value="0.5" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/temp_activity_data_testing.xml" value="0.36822916666666666" />
-        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/temp_data_testing_row.xml" value="0.36822916666666666" />
+        <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/layout/item_scholarship.xml" value="0.536" />
         <entry key="..\:/AndroidStudioProjects/AhCha-Scholarship/AhchaScholarship/app/src/main/res/menu/bottom_nav_menu.xml" value="0.36822916666666666" />
-        <entry key="app/src/main/res/drawable/bottom_border.xml" value="0.3135" />
         <entry key="app/src/main/res/layout/activity_detail_scholarship.xml" value="0.13858695652173914" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.2796875" />
-        <entry key="app/src/main/res/layout/activity_outdoor_filter.xml" value="0.5145216126372849" />
-        <entry key="app/src/main/res/layout/activity_scholarship_filter.xml" value="0.47274457406306614" />
+        <entry key="app/src/main/res/layout/activity_outdoor_filter.xml" value="0.20555555555555555" />
         <entry key="app/src/main/res/layout/favorite.xml" value="0.19375" />
         <entry key="app/src/main/res/layout/fragment_favorite.xml" value="0.346875" />
         <entry key="app/src/main/res/layout/fragment_outdoor.xml" value="0.2513236446657043" />
-        <entry key="app/src/main/res/layout/fragment_scholarship.xml" value="0.3280776977539062" />
+        <entry key="app/src/main/res/layout/fragment_scholarship.xml" value="0.2796875" />
+        <entry key="app/src/main/res/layout/item_outdoor.xml" value="0.20555555555555555" />
         <entry key="app/src/main/res/layout/item_scholarship.xml" value="0.48007246376811596" />
         <entry key="app/src/main/res/layout/temp_activity_data_testing.xml" value="0.28031768798828127" />
         <entry key="app/src/main/res/layout/temp_data_testing_row.xml" value="0.3047225952148439" />
